refactor(kernel): extract mount-table primitives to kernel/mount.rs
Phase G's missing 6th submodule.  Plan G.2 listed 8 submodules
(io / mount / dispatch / ipc / federation / agents / locks /
observability); 5 landed in earlier commits, federation moved out
of kernel entirely in Phase H, agents has no kernel-side methods to
extract — leaving only mount.rs.

  - rust/kernel/src/kernel/mount.rs (new): impl Kernel block over the
    mount-table primitives — add_mount, remove_mount,
    install_mount_metastore, canonical_mount_key, route, has_mount,
    get_mount_points.  Module doc clarifies the layering: these are
    *primitives* composed by `dlc::DriverLifecycleCoordinator` under
    the `sys_setattr(DT_MOUNT)` syscall — not syscalls themselves.
  - rust/kernel/src/kernel/mod.rs: drops the corresponding methods,
    leaves a "moved to kernel::mount" breadcrumb at the
    "// ── Router proxy methods" divider, declares `mod mount;`
    alongside the other syscall-family submodules.

Phase G's submodule split is now complete:

  - kernel/mod.rs      — Kernel struct + constructor + DCache/Router proxies
  - kernel/io.rs       — sys_read/write/stat/unlink/rename/copy/mkdir/rmdir
  - kernel/mount.rs    — mount-table primitives (this commit)
  - kernel/ipc.rs      — pipe + stream registries
  - kernel/locks.rs    — sys_lock / sys_unlock / metastore_list_locks
  - kernel/dispatch.rs — native INTERCEPT hook dispatch
  - kernel/observability.rs — observers + file watches + sys_watch

cargo check + clippy + fmt: clean.
